+++ source/ci_source/providers/Jenkins.ts
@@ -14,6 +14,10 @@ import { pullRequestParser } from "../../platforms/pullRequestParser"
  * [GitHub pull request builder plugin](https://wiki.jenkins.io/display/JENKINS/GitHub+pull+request+builder+plugin)
  * in order to ensure that you have the build environment set up for PR integration.
  *
+ * ### GitLab
+ * You will want to be using the [GitLab Plugin](https://github.com/jenkinsci/gitlab-plugin)
+ * in order to ensure that you have the build environment set up for MR integration.
+ *
  * ### BitBucket Server
  * If using Bitbucket Server, ensure you are using Multibranch Pipelines or Organization Folders.
  * Danger will respect the `CHANGE_URL` and `CHANGE_ID` environment variables.
@@ -50,14 +54,16 @@ export class Jenkins implements CISource {
     const isGitHubPR =
       ensureEnvKeysExist(this.env, ["ghprbPullId", "ghprbGhRepository"]) &&
       ensureEnvKeysAreInt(this.env, ["ghprbPullId"])
+    const isGitLabMR =
+      ensureEnvKeysExist(this.env, ["gitlabMergeRequestIid", "gitlabMergeRequestId"])
     const isMultiBranchPR =
       ensureEnvKeysExist(this.env, ["CHANGE_ID", "CHANGE_URL"]) && ensureEnvKeysAreInt(this.env, ["CHANGE_ID"])
 
-    return this.isJenkins() && (isMultiBranchPR || isGitHubPR)
+    return this.isJenkins() && (isMultiBranchPR || isGitHubPR || isGitLabMR)
   }
 
   get pullRequestID(): string {
-    return this.env.ghprbPullId || this.env.CHANGE_ID
+    return this.env.ghprbPullId || this.env.gitlabMergeRequestIid || this.env.CHANGE_ID
   }
 
   get repoSlug(): string {
@@ -69,7 +75,7 @@ export class Jenkins implements CISource {
 
       return result.repo
     }
-    return this.env.ghprbGhRepository
+    return this.env.ghprbGhRepository || this.env.gitlabTargetNamespace + "/" + this.env.gitlabTargetRepoName
   }
 
   get ciRunURL() {